Use `instance_eval` instead of `eval`
`eval` evaluates code in a global context (which required global `run` and `use`
methods). `instance_eval` evaluates it in the context of an instance.

Previously, Mironfile took in a block that called `eval` on some Ruby code, then
`instance_eval`-ed that block.

Now it takes the path to the file containing that Ruby code and calls
`instance_eval` on the Ruby code directly, without ever calling `eval`.

This means we can get rid of the global `run`/`use` methods.

This PR also removes the code that turned `Miron::Mironfile::CoolClass` into
`CoolClass`, because it's now unnecessary (I think). With `instance_eval` it
should all Just Work.